% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oumuamua.R
\name{oumua.control}
\alias{oumua.control}
\title{Auxiliary for Controlling MARS Estimation}
\usage{
oumua.control(lambda = 1e-08, endspan = NA_integer_,
  minspan = NA_integer_, degree = 1L, nk = 20L, penalty = if
  (degree > 1) 3 else 2, trace = 0L, thresh = 0.001, n_threads = 1,
  K = 1000L)
}
\arguments{
\item{lambda}{L2 penalty mentioned in Friedman (1991, 32). Included both
for numerical stability and potentially better generalization.}

\item{endspan, minspan}{distance between first (last) observation and
respectively the first (last) knot and "internal" knots. See Friedman
(1991, 26-28).}

\item{degree}{maximum degree of basis terms. One yields an additive model.}

\item{nk}{maximum number of basis functions.}

\item{penalty}{penalty in generalized cross validation used in both the
forward and backward pass. See Friedman (1991, 19-22).}

\item{trace}{integer controlling amount of information that is printed to
the console during estimation. Zero yields no information.}

\item{thresh}{required improvement in \eqn{R^2} during forward pass to take
an additional iteration.}

\item{n_threads}{integer with number of threads to use.}

\item{K}{integer greater than zero for the number of basis function to
include before using a queue as suggested by Friedman (1993).}
}
\description{
Auxiliary function for \code{\link{oumua}} and \code{\link{oumua.fit}}.
}
\examples{
str(oumua.control())
str(oumua.control(n_threads = 6L))

}
\references{
Friedman, Jerome H. (1991) \emph{Multivariate Adaptive Regression Splines}.
The Annals of Statistics 19.1: 1-67.

Friedman, Jerome H. (1993) \emph{Fast MARS}.
Stanford University Department of Statistics, Technical Report 110.
}
