% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oumuamua.R
\name{oumua}
\alias{oumua}
\alias{oumua.fit}
\title{Multivariate Adaptive Regression Splines Estimation}
\usage{
oumua(formula, data, offset, control = oumua.control())

oumua.fit(x, y, offset = NULL, do_check = TRUE,
  control = oumua.control())
}
\arguments{
\item{formula}{a \code{\link{formula}} object with responses on the
left-hand-side of \code{~} and covariates on the right-hand-side of
\code{~}.}

\item{data}{\code{\link{environment}} or \code{\link{data.frame}} with
variables in \code{formula}.}

\item{offset}{a symbol or vector with a priori known component of the
linear predictor.}

\item{control}{\code{\link{list}} with arguments passed to
\code{\link{oumua.control}}.}

\item{x}{matrix with covariates.}

\item{y}{observed outcomes.}

\item{do_check}{logical for whether to perform checks of input arguments.}
}
\value{
\code{oumua} returns an object of class oumua. The elements of the
returned object are
\item{X}{design matrix in final model.}
\item{n_vars}{number of basis functions in final model.}
\item{X_scales}{standard deviations of original covariates.}
\item{X_means}{means of original covariates.}
\item{drop_order}{order that basis functions are dropped in during the
backward pass.}
\item{backward_stats}{\eqn{R^2}s and generalized cross validation criteria
during backward pass.}
\item{root_childrens}{objects containing information about the estimated
tree.}
\item{Y}{observed outcomes used in estimation.}
\item{coefficients}{estimated coefficients.}
\item{all_nodes}{"unlisted" version of \code{root_childrens}.}
\item{call}{matched call.}
\item{terms}{\code{\link{terms.object}} from \code{\link{model.frame}}
call.}
}
\description{
Runs the Multivariate Adaptive Regression Splines (MARS)
algorithm potentially in parallel.
}
\details{
Runs the MARS algorithm as suggested in Friedman (1991). The README at
\url{https://github.com/boennecd/Oumuamua} contains simulations examples
which also illustrates how to use the functions in the package.
}
\examples{
library(Oumuamua)
data("mtcars")
f1 <- oumua(mpg ~ cyl + disp + hp + drat + wt, mtcars,
            control = oumua.control(lambda = 1, endspan = 1L, minspan = 1L,
                                    degree = 2, n_threads = 1))
f2 <- oumua(mpg ~ cyl + disp + hp + drat + wt, mtcars,
            control = oumua.control(lambda = 1, endspan = 1L, minspan = 1L,
                                    degree = 2, n_threads = 2))
stopifnot(all.equal(coef(f1), coef(f2)))

}
\references{
Friedman, Jerome H. \emph{Multivariate Adaptive Regression Splines}.
The Annals of Statistics 19.1 (1991): 1-67.
}
