% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oumuamua.R
\name{predict.oumua}
\alias{predict.oumua}
\title{Predict Method for MARS Fits}
\usage{
\method{predict}{oumua}(object, newdata, ...)
}
\arguments{
\item{object}{object of class \code{oumua}.}

\item{newdata}{\code{\link{data.frame}} with new observations.}

\item{...}{un-used.}
}
\description{
Returns predicted values for new observations form an estimated MARS
model.
}
\examples{
library(Oumuamua)
data("mtcars")
fit <- oumua(mpg ~ cyl + disp + hp + drat + wt, mtcars,
             control = oumua.control(lambda = 1, endspan = 1L, minspan = 1L,
                                     degree = 2))
stopifnot(all.equal(drop(fit$X[1:10, ] \%*\% coef(fit)),
                    predict(fit, mtcars[1:10, ])))

}
