% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oumuamua.R
\name{summary.oumua}
\alias{summary.oumua}
\title{Summarizing MARS Fits}
\usage{
\method{summary}{oumua}(object, ...)
}
\arguments{
\item{object}{object of class \code{oumua}.}

\item{...}{un-used.}
}
\description{
Adds the \eqn{R^2} to the \code{oumua} object which can then be used in
\code{print.summary.oumua}.
}
\examples{
library(Oumuamua)
data("mtcars")
fit <- oumua(mpg ~ cyl + disp + hp + drat + wt, mtcars,
             control = oumua.control(lambda = 1, endspan = 1L, minspan = 1L,
                                     degree = 2))
summary(fit)

}
